# server/docs/swagger.json
{
"openapi": "3.1.0",
"info": {
        "contact": {
            "name": "Rutoso project",
            "url": "https://github.com/achetronic/rutoso"
        },
        "description": "REST API control plane for Envoy proxies. Manage route groups and routes;\nchanges are pushed to all connected Envoy instances via xDS in real time.",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "title": "Rutoso API",
        "version": "1.0"
    },
"components": {
        "schemas": {
...
            "model.CORSConfig": {
                "description": "CORS holds the CORS filter configuration. Set when Type == \"cors\".",
                "properties": {
                    "allowCredentials": {
                        "description": "AllowCredentials indicates whether the request can include user credentials.",
                        "type": "boolean"
                    },
                    "allowHeaders": {
                        "description": "AllowHeaders lists the request headers allowed in CORS requests.",
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "allowMethods": {
                        "description": "AllowMethods lists the HTTP methods allowed in CORS requests.\nExample: [\"GET\", \"POST\", \"OPTIONS\"]",
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "allowOrigins": {
                        "description": "AllowOrigins lists the allowed origin patterns.\nEach entry is matched as an exact string or a regex if Regex is true.",
                        "items": {
                            "$ref": "#/components/schemas/model.CORSOrigin"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "exposeHeaders": {
                        "description": "ExposeHeaders lists the response headers the browser is allowed to access.",
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "maxAge": {
                        "description": "MaxAge sets the preflight cache duration in seconds.\nMaps to the Access-Control-Max-Age response header.",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "model.CORSOrigin": {
                "properties": {
                    "regex": {
                        "description": "Regex indicates that Value should be treated as a regular expression.",
                        "type": "boolean"
                    },
                    "value": {
                        "description": "Value is the origin string or regex pattern.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "model.ExtAuthzConfig": {
                "description": "ExtAuthz holds the external authorisation filter configuration.\nSet when Type == \"extAuthz\".",
                "properties": {
                    "failureModeAllow": {
                        "description": "FailureModeAllow controls what happens when the authz service is\nunreachable. If true, requests are allowed through (fail-open).\nDefault is false (fail-closed).",
                        "type": "boolean"
                    },
                    "grpcService": {
                        "description": "GRPCService is the address of the gRPC authorisation service\n(e.g. \"authz.default.svc.cluster.local:50051\").\nMutually exclusive with HTTPService.",
                        "type": "string"
                    },
                    "httpService": {
                        "description": "HTTPService is the URL of the HTTP authorisation service\n(e.g. \"http://opa.default.svc.cluster.local:8181/v1/authz\").\nMutually exclusive with GRPCService.",
                        "type": "string"
                    },
                    "includeRequestBodyInCheck": {
                        "description": "IncludeRequestBodyInCheck forwards the request body to the authz service.",
                        "type": "boolean"
                    },
                    "timeout": {
                        "description": "Timeout is the authorisation request deadline (e.g. \"5s\", \"500ms\").",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "model.ExtProcConfig": {
                "description": "ExtProc holds the external processing filter configuration.\nSet when Type == \"extProc\".",
                "properties": {
                    "grpcService": {
                        "description": "GRPCService is the address of the gRPC processing service\n(e.g. \"ext-proc.default.svc.cluster.local:9000\").",
                        "type": "string"
                    },
                    "processingMode": {
                        "$ref": "#/components/schemas/model.ExtProcMode"
                    },
                    "timeout": {
                        "description": "Timeout is the processing request deadline (e.g. \"2s\").",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "model.ExtProcMode": {
                "description": "ProcessingMode controls which parts of the HTTP transaction are sent to\nthe external processor.",
                "properties": {
                    "requestBodyMode": {
                        "$ref": "#/components/schemas/model.ExtProcPhase"
                    },
                    "requestHeaderMode": {
                        "$ref": "#/components/schemas/model.ExtProcPhase"
                    },
                    "responseBodyMode": {
                        "$ref": "#/components/schemas/model.ExtProcPhase"
                    },
                    "responseHeaderMode": {
                        "$ref": "#/components/schemas/model.ExtProcPhase"
                    }
                },
                "type": "object"
            },
            "model.ExtProcPhase": {
                "description": "ResponseBodyMode controls processing of the response body.",
                "enum": [
                    "SKIP",
                    "SEND",
                    "BUFFERED"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "ExtProcPhaseSkip",
                    "ExtProcPhaseSend",
                    "ExtProcPhaseBuffered"
                ]
            },
...
            "model.HeaderValue": {
                "properties": {
                    "append": {
                        "description": "Append controls whether the header is appended (true) or replaced\n(false) if it already exists. Default: true.",
                        "type": "boolean"
                    },
                    "key": {
                        "description": "Key is the header name.",
                        "type": "string"
                    },
                    "value": {
                        "description": "Value is the header value.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "model.HeadersConfig": {
                "description": "Headers holds the header mutation filter configuration.\nSet when Type == \"headers\".",
                "properties": {
                    "requestHeadersToAdd": {
                        "description": "RequestHeadersToAdd are headers added to the request before forwarding\nto the upstream.",
                        "items": {
                            "$ref": "#/components/schemas/model.HeaderValue"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "requestHeadersToRemove": {
                        "description": "RequestHeadersToRemove are header names removed from the request\nbefore forwarding.",
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "responseHeadersToAdd": {
                        "description": "ResponseHeadersToAdd are headers added to the response before\nreturning to the client.",
                        "items": {
                            "$ref": "#/components/schemas/model.HeaderValue"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "responseHeadersToRemove": {
                        "description": "ResponseHeadersToRemove are header names removed from the response\nbefore returning to the client.",
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
...
            "model.JWTClaimHeader": {
                "properties": {
                    "claim": {
                        "description": "Claim is the JWT claim name (e.g. \"sub\", \"email\").",
                        "type": "string"
                    },
                    "header": {
                        "description": "Header is the request header name that receives the claim value.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "model.JWTConfig": {
                "description": "JWT holds the JWT authn filter configuration. Set when Type == \"jwt\".",
                "properties": {
                    "providers": {
                        "additionalProperties": {
                            "$ref": "#/components/schemas/model.JWTProvider"
                        },
                        "description": "Providers is a map of provider name to JWTProvider configuration.\nThe map key is referenced from per-route overrides to select a specific\nprovider (or disable authentication for that route).",
                        "type": "object"
                    },
                    "rules": {
                        "description": "Rules defines which request paths require JWT validation and which\nprovider to apply. If empty, all paths are validated by all providers.",
                        "items": {
                            "$ref": "#/components/schemas/model.JWTRule"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
            "model.JWTProvider": {
                "properties": {
                    "audiences": {
                        "description": "Audiences lists the expected \"aud\" claim values. If empty, audience\nvalidation is skipped.",
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "claimToHeaders": {
                        "description": "ClaimToHeaders maps JWT claim names to upstream request header names.\nThe claim value is set as the header value before forwarding.",
                        "items": {
                            "$ref": "#/components/schemas/model.JWTClaimHeader"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "forwardJwt": {
                        "description": "ForwardJWT indicates whether the original Authorization header should be\nforwarded to the upstream after successful validation.",
                        "type": "boolean"
                    },
                    "issuer": {
                        "description": "Issuer is the expected value of the \"iss\" claim.",
                        "type": "string"
                    },
                    "jwksInline": {
                        "description": "JWKsInline is a literal JSON Web Key Set document.\nMutually exclusive with JWKsURI.",
                        "type": "string"
                    },
                    "jwksUri": {
                        "description": "JWKsURI is the URL from which the JSON Web Key Set is fetched.\nMutually exclusive with JWKsInline.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "model.JWTRule": {
                "properties": {
                    "allowMissing": {
                        "description": "AllowMissing allows requests without a JWT token to pass through.\nUseful for public endpoints that coexist in the same listener.",
                        "type": "boolean"
                    },
                    "match": {
                        "description": "Match is the path prefix this rule applies to.",
                        "type": "string"
                    },
                    "requires": {
                        "description": "Requires lists the provider names that must validate the request.\nAll listed providers must succeed (AND semantics).",
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
...
            "model.Middleware": {
                "properties": {
                    "cors": {
                        "$ref": "#/components/schemas/model.CORSConfig"
                    },
                    "extAuthz": {
                        "$ref": "#/components/schemas/model.ExtAuthzConfig"
                    },
                    "extProc": {
                        "$ref": "#/components/schemas/model.ExtProcConfig"
                    },
                    "headers": {
                        "$ref": "#/components/schemas/model.HeadersConfig"
                    },
                    "id": {
                        "description": "ID is the unique identifier of the filter.",
                        "type": "string"
                    },
                    "jwt": {
                        "$ref": "#/components/schemas/model.JWTConfig"
                    },
                    "name": {
                        "description": "Name is a human-readable label for the filter.",
                        "type": "string"
                    },
                    "rateLimit": {
                        "$ref": "#/components/schemas/model.RateLimitConfig"
                    },
                    "type": {
                        "$ref": "#/components/schemas/model.MiddlewareType"
                    }
                },
                "type": "object"
            },
...
            "model.MiddlewareType": {
                "description": "Type identifies which Envoy HTTP filter this entity configures.\nExactly one of the Config* fields below must match this type.",
                "enum": [
                    "cors",
                    "jwt",
                    "extAuthz",
                    "extProc",
                    "rateLimit",
                    "headers"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "MiddlewareTypeCORS",
                    "MiddlewareTypeJWT",
                    "MiddlewareTypeExtAuthz",
                    "MiddlewareTypeExtProc",
                    "MiddlewareTypeRateLimit",
                    "MiddlewareTypeHeaders"
                ]
            },
...
            "model.RateLimitConfig": {
                "description": "RateLimit holds the rate limit filter configuration.\nSet when Type == \"rateLimit\".",
                "properties": {
                    "domain": {
                        "description": "Domain is the rate limit domain passed to the rate limit service.\nThe service uses this to scope rate limit descriptors.",
                        "type": "string"
                    },
                    "failureModeDeny": {
                        "description": "FailureModeDeny controls what happens when the rate limit service is\nunreachable. If true, requests are denied (fail-closed).\nDefault is false (fail-open).",
                        "type": "boolean"
                    },
                    "grpcService": {
                        "description": "GRPCService is the address of the gRPC rate limit service\n(e.g. \"ratelimit.default.svc.cluster.local:8081\").",
                        "type": "string"
                    },
                    "timeout": {
                        "description": "Timeout is the rate limit request deadline (e.g. \"500ms\").",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "respond.ErrorBody": {
                "properties": {
                    "error": {
                        "description": "Error contains a human-readable description of what went wrong.",
                        "example": "group \"abc\" not found",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
...
"paths": {
...
        "/middlewares": {
            "get": {
                "description": "Returns the full list of middlewares.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/model.Middleware"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/respond.ErrorBody"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "List middlewares",
                "tags": [
                    "middlewares"
                ]
            },
            "post": {
                "description": "Creates a new HTTP mw entity.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/model.Middleware",
                                        "summary": "mw",
                                        "description": "Filter definition"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Filter definition",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/model.Middleware"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/respond.ErrorBody"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/respond.ErrorBody"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Create a middleware",
                "tags": [
                    "middlewares"
                ]
            }
        },
        "/middlewares/{middlewareId}": {
            "delete": {
                "description": "Deletes the mw with the given ID.",
                "parameters": [
                    {
                        "description": "Filter ID",
                        "in": "path",
                        "name": "middlewareId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/respond.ErrorBody"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/respond.ErrorBody"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Delete a middleware",
                "tags": [
                    "middlewares"
                ]
            },
            "get": {
                "description": "Returns the mw with the given ID.",
                "parameters": [
                    {
                        "description": "Filter ID",
                        "in": "path",
                        "name": "middlewareId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/model.Middleware"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/respond.ErrorBody"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/respond.ErrorBody"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Get a middleware",
                "tags": [
                    "middlewares"
                ]
            },
            "put": {
                "description": "Replaces the mw with the given ID.",
                "parameters": [
                    {
                        "description": "Filter ID",
                        "in": "path",
                        "name": "middlewareId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/model.Middleware",
                                        "summary": "mw",
                                        "description": "Updated mw definition"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Updated mw definition",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/model.Middleware"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/respond.ErrorBody"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/respond.ErrorBody"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/respond.ErrorBody"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Update a middleware",
                "tags": [
                    "middlewares"
                ]
            }
        },
...
    },
"servers": [
        {
            "url": "localhost:8080/"
        }
    ]
}